CAMEL-24003: Extract TuiToolRegistry to share tools between AiPanel and TuiMcpServer
